+++ lib/ductwork/processes/supervisor_runner.rb
@@ -3,32 +3,37 @@
 module Ductwork
   module Processes
     class SupervisorRunner
-      def self.start!
-        supervisor = Ductwork::Processes::Supervisor.new
-        pipelines_to_advance = Ductwork.configuration.pipelines
+      def initialize(*pipelines)
+        @pipelines = pipelines
+        @supervisor = Ductwork::Processes::Supervisor.new
+      end
 
-        supervisor.add_worker(metadata: { pipelines: pipelines_to_advance }) do
+      def run
+        supervisor.add_worker(metadata: { pipelines: }) do
           Ductwork.logger.debug(
             msg: "Starting Pipeline Advancer process",
             role: :supervisor_runner
           )
-          Ductwork::Processes::PipelineAdvancerRunner
-            .new(*pipelines_to_advance).run
+          Ductwork::Processes::PipelineAdvancerRunner.new(*pipelines).run
         end
 
-        pipelines_to_advance.each do |pipeline|
+        pipelines.each do |pipeline|
           supervisor.add_worker(metadata: { pipeline: }) do
             Ductwork.logger.debug(
               msg: "Starting Job Worker Runner process",
               role: :supervisor_runner,
               pipeline: pipeline
             )
-            Ductwork::Processes::JobWorkerRunner.new(pipeline).run
+            Ductwork::Processes::JobWorkerRunner.new(*pipeline).run
           end
         end
 
         supervisor.run
       end
+
+      private
+
+      attr_reader :pipelines, :supervisor
     end
   end
 end