+++ spec/ductwork/cli_spec.rb
@@ -13,7 +13,7 @@ RSpec.describe Ductwork::CLI do
 
     before do
       ENV.delete("DUCTWORK_ROLE")
-      allow(Ductwork::Processes::SupervisorRunner).to receive(:start!)
+      allow(Ductwork::Processes::Launcher).to receive(:start_processes!)
       allow(Ductwork::Configuration).to receive(:new).and_return(config)
       allow(Ductwork).to receive(:logger=).and_call_original
       allow(Ductwork).to receive(:logger).and_return(logger)
@@ -28,13 +28,11 @@ RSpec.describe Ductwork::CLI do
     end
 
     it "loads the role from ENV" do
-      ENV["DUCTWORK_ROLE"] = "supervisor"
+      ENV["DUCTWORK_ROLE"] = "advancer"
 
       described_class.start!([])
 
-      expect(Ductwork::Configuration).to have_received(:new).with(
-        role: "supervisor"
-      )
+      expect(Ductwork::Configuration).to have_received(:new).with(role: "advancer")
     end
 
     it "prints the banner" do
@@ -57,10 +55,10 @@ RSpec.describe Ductwork::CLI do
       BANNER
     end
 
-    it "starts the supervisor" do
+    it "calls the process launcher" do
       described_class.start!([])
 
-      expect(Ductwork::Processes::SupervisorRunner).to have_received(:start!)
+      expect(Ductwork::Processes::Launcher).to have_received(:start_processes!)
     end
   end
 end